requirements.txt
fastapi==0.135.1
uvicorn==0.42.0
langchain==1.2.13
langchain-openai==1.1.11
langchain-community==0.4.1
langgraph==1.1.3
chromadb==1.5.5
transformers==5.3.0
torch==2.10.0
prophet==1.3.0
pandas==3.0.1
numpy==2.4.3
scipy==1.17.1
python-dotenv==1.2.2
httpx==0.28.1
newsapi-python==0.2.7
pydantic==2.12.5
reportlab==4.4.10
